refactor: split WordDocument part generators
