+++ OfficeIMO.Word/Parts/WordDocument.Parts.DocDefaults.cs
@@ -1,0 +1,62 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Text;
+using DocumentFormat.OpenXml;
+using DocumentFormat.OpenXml.ExtendedProperties;
+using DocumentFormat.OpenXml.Packaging;
+using DocumentFormat.OpenXml.Wordprocessing;
+using Ap = DocumentFormat.OpenXml.ExtendedProperties;
+using Vt = DocumentFormat.OpenXml.VariantTypes;
+using M = DocumentFormat.OpenXml.Math;
+using Ovml = DocumentFormat.OpenXml.Vml.Office;
+using V = DocumentFormat.OpenXml.Vml;
+using W14 = DocumentFormat.OpenXml.Office2010.Word;
+using W15 = DocumentFormat.OpenXml.Office2013.Word;
+using A = DocumentFormat.OpenXml.Drawing;
+using Thm15 = DocumentFormat.OpenXml.Office2013.Theme;
+using DocumentFormat.OpenXml.Drawing.Diagrams;
+using Style = DocumentFormat.OpenXml.Wordprocessing.Style;
+
+namespace OfficeIMO.Word {
+    public partial class WordDocument {
+        /// <summary>
+        /// Defaults for Word Document
+        /// </summary>
+        /// <returns></returns>
+        private static DocDefaults GenerateDocDefaults() {
+            DocDefaults docDefaults1 = new DocDefaults();
+
+            RunPropertiesDefault runPropertiesDefault1 = new RunPropertiesDefault();
+
+            RunPropertiesBaseStyle runPropertiesBaseStyle1 = new RunPropertiesBaseStyle();
+            RunFonts runFonts1 = new RunFonts() { AsciiTheme = ThemeFontValues.MinorHighAnsi, HighAnsiTheme = ThemeFontValues.MinorHighAnsi, EastAsiaTheme = ThemeFontValues.MinorHighAnsi, ComplexScriptTheme = ThemeFontValues.MinorBidi };
+            FontSize fontSize1 = new FontSize() { Val = "22" };
+            FontSizeComplexScript fontSizeComplexScript1 = new FontSizeComplexScript() { Val = "22" };
+            Languages languages1 = new Languages() { Val = "en-US", EastAsia = "en-US", Bidi = "ar-SA" };
+            //Languages languages1 = new Languages() { Val = "pl-PL", EastAsia = "pl-PL", Bidi = "ar-SA" };
+
+            runPropertiesBaseStyle1.Append(runFonts1);
+            runPropertiesBaseStyle1.Append(fontSize1);
+            runPropertiesBaseStyle1.Append(fontSizeComplexScript1);
+            runPropertiesBaseStyle1.Append(languages1);
+
+            runPropertiesDefault1.Append(runPropertiesBaseStyle1);
+
+            ParagraphPropertiesDefault paragraphPropertiesDefault1 = new ParagraphPropertiesDefault();
+
+            ParagraphPropertiesBaseStyle paragraphPropertiesBaseStyle1 = new ParagraphPropertiesBaseStyle();
+            SpacingBetweenLines spacingBetweenLines1 = new SpacingBetweenLines() { After = "160", Line = "259", LineRule = LineSpacingRuleValues.Auto };
+
+            paragraphPropertiesBaseStyle1.Append(spacingBetweenLines1);
+
+            paragraphPropertiesDefault1.Append(paragraphPropertiesBaseStyle1);
+
+            docDefaults1.Append(runPropertiesDefault1);
+            docDefaults1.Append(paragraphPropertiesDefault1);
+            return docDefaults1;
+        }
+    }
+}
