+++ OfficeIMO.Word/Parts/WordDocument.Parts.FontTable.cs
@@ -1,0 +1,85 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Text;
+using DocumentFormat.OpenXml;
+using DocumentFormat.OpenXml.ExtendedProperties;
+using DocumentFormat.OpenXml.Packaging;
+using DocumentFormat.OpenXml.Wordprocessing;
+using Ap = DocumentFormat.OpenXml.ExtendedProperties;
+using Vt = DocumentFormat.OpenXml.VariantTypes;
+using M = DocumentFormat.OpenXml.Math;
+using Ovml = DocumentFormat.OpenXml.Vml.Office;
+using V = DocumentFormat.OpenXml.Vml;
+using W14 = DocumentFormat.OpenXml.Office2010.Word;
+using W15 = DocumentFormat.OpenXml.Office2013.Word;
+using A = DocumentFormat.OpenXml.Drawing;
+using Thm15 = DocumentFormat.OpenXml.Office2013.Theme;
+using DocumentFormat.OpenXml.Drawing.Diagrams;
+using Style = DocumentFormat.OpenXml.Wordprocessing.Style;
+
+namespace OfficeIMO.Word {
+    public partial class WordDocument {
+        // Generates content of fontTablePart1.
+        private static void GenerateFontTablePart1Content(FontTablePart fontTablePart1) {
+            Fonts fonts1 = new Fonts() { MCAttributes = new MarkupCompatibilityAttributes() { Ignorable = "w14 w15 w16se w16cid w16 w16cex w16sdtdh" } };
+            fonts1.AddNamespaceDeclaration("mc", "http://schemas.openxmlformats.org/markup-compatibility/2006");
+            fonts1.AddNamespaceDeclaration("r", "http://schemas.openxmlformats.org/officeDocument/2006/relationships");
+            fonts1.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
+            fonts1.AddNamespaceDeclaration("w14", "http://schemas.microsoft.com/office/word/2010/wordml");
+            fonts1.AddNamespaceDeclaration("w15", "http://schemas.microsoft.com/office/word/2012/wordml");
+            fonts1.AddNamespaceDeclaration("w16cex", "http://schemas.microsoft.com/office/word/2018/wordml/cex");
+            fonts1.AddNamespaceDeclaration("w16cid", "http://schemas.microsoft.com/office/word/2016/wordml/cid");
+            fonts1.AddNamespaceDeclaration("w16", "http://schemas.microsoft.com/office/word/2018/wordml");
+            fonts1.AddNamespaceDeclaration("w16sdtdh", "http://schemas.microsoft.com/office/word/2020/wordml/sdtdatahash");
+            fonts1.AddNamespaceDeclaration("w16se", "http://schemas.microsoft.com/office/word/2015/wordml/symex");
+
+            Font font1 = new Font() { Name = "Calibri" };
+            Panose1Number panose1Number1 = new Panose1Number() { Val = "020F0502020204030204" };
+            FontCharSet fontCharSet1 = new FontCharSet() { Val = "00" };
+            FontFamily fontFamily1 = new FontFamily() { Val = FontFamilyValues.Swiss };
+            Pitch pitch1 = new Pitch() { Val = FontPitchValues.Variable };
+            FontSignature fontSignature1 = new FontSignature() { UnicodeSignature0 = "E4002EFF", UnicodeSignature1 = "C000247B", UnicodeSignature2 = "00000009", UnicodeSignature3 = "00000000", CodePageSignature0 = "000001FF", CodePageSignature1 = "00000000" };
+
+            font1.Append(panose1Number1);
+            font1.Append(fontCharSet1);
+            font1.Append(fontFamily1);
+            font1.Append(pitch1);
+            font1.Append(fontSignature1);
+
+            Font font2 = new Font() { Name = "Times New Roman" };
+            Panose1Number panose1Number2 = new Panose1Number() { Val = "02020603050405020304" };
+            FontCharSet fontCharSet2 = new FontCharSet() { Val = "00" };
+            FontFamily fontFamily2 = new FontFamily() { Val = FontFamilyValues.Roman };
+            Pitch pitch2 = new Pitch() { Val = FontPitchValues.Variable };
+            FontSignature fontSignature2 = new FontSignature() { UnicodeSignature0 = "E0002EFF", UnicodeSignature1 = "C000785B", UnicodeSignature2 = "00000009", UnicodeSignature3 = "00000000", CodePageSignature0 = "000001FF", CodePageSignature1 = "00000000" };
+
+            font2.Append(panose1Number2);
+            font2.Append(fontCharSet2);
+            font2.Append(fontFamily2);
+            font2.Append(pitch2);
+            font2.Append(fontSignature2);
+
+            Font font3 = new Font() { Name = "Calibri Light" };
+            Panose1Number panose1Number3 = new Panose1Number() { Val = "020F0302020204030204" };
+            FontCharSet fontCharSet3 = new FontCharSet() { Val = "00" };
+            FontFamily fontFamily3 = new FontFamily() { Val = FontFamilyValues.Swiss };
+            Pitch pitch3 = new Pitch() { Val = FontPitchValues.Variable };
+            FontSignature fontSignature3 = new FontSignature() { UnicodeSignature0 = "E4002EFF", UnicodeSignature1 = "C000247B", UnicodeSignature2 = "00000009", UnicodeSignature3 = "00000000", CodePageSignature0 = "000001FF", CodePageSignature1 = "00000000" };
+
+            font3.Append(panose1Number3);
+            font3.Append(fontCharSet3);
+            font3.Append(fontFamily3);
+            font3.Append(pitch3);
+            font3.Append(fontSignature3);
+
+            fonts1.Append(font1);
+            fonts1.Append(font2);
+            fonts1.Append(font3);
+
+            fontTablePart1.Fonts = fonts1;
+        }
+    }
+}
