+++ OfficeIMO.Word/Parts/WordDocument.Parts.Settings.cs
@@ -1,0 +1,174 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Text;
+using DocumentFormat.OpenXml;
+using DocumentFormat.OpenXml.ExtendedProperties;
+using DocumentFormat.OpenXml.Packaging;
+using DocumentFormat.OpenXml.Wordprocessing;
+using Ap = DocumentFormat.OpenXml.ExtendedProperties;
+using Vt = DocumentFormat.OpenXml.VariantTypes;
+using M = DocumentFormat.OpenXml.Math;
+using Ovml = DocumentFormat.OpenXml.Vml.Office;
+using V = DocumentFormat.OpenXml.Vml;
+using W14 = DocumentFormat.OpenXml.Office2010.Word;
+using W15 = DocumentFormat.OpenXml.Office2013.Word;
+using A = DocumentFormat.OpenXml.Drawing;
+using Thm15 = DocumentFormat.OpenXml.Office2013.Theme;
+using DocumentFormat.OpenXml.Drawing.Diagrams;
+using Style = DocumentFormat.OpenXml.Wordprocessing.Style;
+
+namespace OfficeIMO.Word {
+    public partial class WordDocument {
+        /// <summary>
+        /// Adds default settings for WebSettings
+        /// This is the default from Microsoft Word in Microsoft Office 365
+        /// </summary>
+        /// <param name="webSettingsPart1"></param>
+        private static void GenerateWebSettingsPart1Content(WebSettingsPart webSettingsPart1) {
+            WebSettings webSettings1 = new WebSettings() { MCAttributes = new MarkupCompatibilityAttributes() { Ignorable = "w14 w15 w16se w16cid w16 w16cex w16sdtdh" } };
+            webSettings1.AddNamespaceDeclaration("mc", "http://schemas.openxmlformats.org/markup-compatibility/2006");
+            webSettings1.AddNamespaceDeclaration("r", "http://schemas.openxmlformats.org/officeDocument/2006/relationships");
+            webSettings1.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
+            webSettings1.AddNamespaceDeclaration("w14", "http://schemas.microsoft.com/office/word/2010/wordml");
+            webSettings1.AddNamespaceDeclaration("w15", "http://schemas.microsoft.com/office/word/2012/wordml");
+            webSettings1.AddNamespaceDeclaration("w16cex", "http://schemas.microsoft.com/office/word/2018/wordml/cex");
+            webSettings1.AddNamespaceDeclaration("w16cid", "http://schemas.microsoft.com/office/word/2016/wordml/cid");
+            webSettings1.AddNamespaceDeclaration("w16", "http://schemas.microsoft.com/office/word/2018/wordml");
+            webSettings1.AddNamespaceDeclaration("w16sdtdh", "http://schemas.microsoft.com/office/word/2020/wordml/sdtdatahash");
+            webSettings1.AddNamespaceDeclaration("w16se", "http://schemas.microsoft.com/office/word/2015/wordml/symex");
+            OptimizeForBrowser optimizeForBrowser1 = new OptimizeForBrowser();
+            AllowPNG allowPNG1 = new AllowPNG();
+
+            webSettings1.Append(optimizeForBrowser1);
+            webSettings1.Append(allowPNG1);
+
+            webSettingsPart1.WebSettings = webSettings1;
+        }
+
+        /// <summary>
+        /// Adds default settings for Word Document
+        /// This is the default from Microsoft Word in Microsoft Office 365
+        /// </summary>
+        /// <param name="documentSettingsPart1"></param>
+        private static void GenerateDocumentSettingsPart1Content(DocumentSettingsPart documentSettingsPart1) {
+            Settings settings1 = new Settings() { MCAttributes = new MarkupCompatibilityAttributes() { Ignorable = "w14 w15 w16se w16cid w16 w16cex w16sdtdh" } };
+            settings1.AddNamespaceDeclaration("mc", "http://schemas.openxmlformats.org/markup-compatibility/2006");
+            settings1.AddNamespaceDeclaration("o", "urn:schemas-microsoft-com:office:office");
+            settings1.AddNamespaceDeclaration("r", "http://schemas.openxmlformats.org/officeDocument/2006/relationships");
+            settings1.AddNamespaceDeclaration("m", "http://schemas.openxmlformats.org/officeDocument/2006/math");
+            settings1.AddNamespaceDeclaration("v", "urn:schemas-microsoft-com:vml");
+            settings1.AddNamespaceDeclaration("w10", "urn:schemas-microsoft-com:office:word");
+            settings1.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
+            settings1.AddNamespaceDeclaration("w14", "http://schemas.microsoft.com/office/word/2010/wordml");
+            settings1.AddNamespaceDeclaration("w15", "http://schemas.microsoft.com/office/word/2012/wordml");
+            settings1.AddNamespaceDeclaration("w16cex", "http://schemas.microsoft.com/office/word/2018/wordml/cex");
+            settings1.AddNamespaceDeclaration("w16cid", "http://schemas.microsoft.com/office/word/2016/wordml/cid");
+            settings1.AddNamespaceDeclaration("w16", "http://schemas.microsoft.com/office/word/2018/wordml");
+            settings1.AddNamespaceDeclaration("w16sdtdh", "http://schemas.microsoft.com/office/word/2020/wordml/sdtdatahash");
+            settings1.AddNamespaceDeclaration("w16se", "http://schemas.microsoft.com/office/word/2015/wordml/symex");
+            settings1.AddNamespaceDeclaration("sl", "http://schemas.openxmlformats.org/schemaLibrary/2006/main");
+            Zoom zoom1 = new Zoom() { Percent = "100" };
+            ProofState proofState1 = new ProofState() { Spelling = ProofingStateValues.Clean, Grammar = ProofingStateValues.Clean };
+            DefaultTabStop defaultTabStop1 = new DefaultTabStop() { Val = 720 };
+            CharacterSpacingControl characterSpacingControl1 = new CharacterSpacingControl() { Val = CharacterSpacingValues.DoNotCompress };
+
+            Compatibility compatibility1 = new Compatibility();
+            CompatibilitySetting compatibilitySetting1 = new CompatibilitySetting() { Name = CompatSettingNameValues.CompatibilityMode, Uri = "http://schemas.microsoft.com/office/word", Val = "15" };
+            CompatibilitySetting compatibilitySetting2 = new CompatibilitySetting() { Name = CompatSettingNameValues.OverrideTableStyleFontSizeAndJustification, Uri = "http://schemas.microsoft.com/office/word", Val = "1" };
+            CompatibilitySetting compatibilitySetting3 = new CompatibilitySetting() { Name = CompatSettingNameValues.EnableOpenTypeFeatures, Uri = "http://schemas.microsoft.com/office/word", Val = "1" };
+            CompatibilitySetting compatibilitySetting4 = new CompatibilitySetting() { Name = CompatSettingNameValues.DoNotFlipMirrorIndents, Uri = "http://schemas.microsoft.com/office/word", Val = "1" };
+            CompatibilitySetting compatibilitySetting5 = new CompatibilitySetting() { Name = CompatSettingNameValues.DifferentiateMultirowTableHeaders, Uri = "http://schemas.microsoft.com/office/word", Val = "1" };
+            CompatibilitySetting compatibilitySetting6 = new CompatibilitySetting() { Name = new EnumValue<CompatSettingNameValues>() { InnerText = "useWord2013TrackBottomHyphenation" }, Uri = "http://schemas.microsoft.com/office/word", Val = "0" };
+
+            compatibility1.Append(compatibilitySetting1);
+            compatibility1.Append(compatibilitySetting2);
+            compatibility1.Append(compatibilitySetting3);
+            compatibility1.Append(compatibilitySetting4);
+            compatibility1.Append(compatibilitySetting5);
+            compatibility1.Append(compatibilitySetting6);
+
+            Rsids rsids1 = new Rsids();
+            RsidRoot rsidRoot1 = new RsidRoot() { Val = "002D4FAA" };
+            Rsid rsid1 = new Rsid() { Val = "002D4FAA" };
+            Rsid rsid2 = new Rsid() { Val = "006B3CD0" };
+            Rsid rsid3 = new Rsid() { Val = "00767C75" };
+            Rsid rsid4 = new Rsid() { Val = "00902D87" };
+            Rsid rsid5 = new Rsid() { Val = "00E41DA3" };
+            Rsid rsid6 = new Rsid() { Val = "00E82D66" };
+            Rsid rsid7 = new Rsid() { Val = "00FF08E3" };
+
+            rsids1.Append(rsidRoot1);
+            rsids1.Append(rsid1);
+            rsids1.Append(rsid2);
+            rsids1.Append(rsid3);
+            rsids1.Append(rsid4);
+            rsids1.Append(rsid5);
+            rsids1.Append(rsid6);
+            rsids1.Append(rsid7);
+
+            M.MathProperties mathProperties1 = new M.MathProperties();
+            M.MathFont mathFont1 = new M.MathFont() { Val = "Cambria Math" };
+            M.BreakBinary breakBinary1 = new M.BreakBinary() { Val = M.BreakBinaryOperatorValues.Before };
+            M.BreakBinarySubtraction breakBinarySubtraction1 = new M.BreakBinarySubtraction() { Val = M.BreakBinarySubtractionValues.MinusMinus };
+            M.SmallFraction smallFraction1 = new M.SmallFraction() { Val = M.BooleanValues.Zero };
+            M.DisplayDefaults displayDefaults1 = new M.DisplayDefaults();
+            M.LeftMargin leftMargin1 = new M.LeftMargin() { Val = (UInt32Value)0U };
+            M.RightMargin rightMargin1 = new M.RightMargin() { Val = (UInt32Value)0U };
+            M.DefaultJustification defaultJustification1 = new M.DefaultJustification() { Val = M.JustificationValues.CenterGroup };
+            M.WrapIndent wrapIndent1 = new M.WrapIndent() { Val = (UInt32Value)1440U };
+            M.IntegralLimitLocation integralLimitLocation1 = new M.IntegralLimitLocation() { Val = M.LimitLocationValues.SubscriptSuperscript };
+            M.NaryLimitLocation naryLimitLocation1 = new M.NaryLimitLocation() { Val = M.LimitLocationValues.UnderOver };
+
+            mathProperties1.Append(mathFont1);
+            mathProperties1.Append(breakBinary1);
+            mathProperties1.Append(breakBinarySubtraction1);
+            mathProperties1.Append(smallFraction1);
+            mathProperties1.Append(displayDefaults1);
+            mathProperties1.Append(leftMargin1);
+            mathProperties1.Append(rightMargin1);
+            mathProperties1.Append(defaultJustification1);
+            mathProperties1.Append(wrapIndent1);
+            mathProperties1.Append(integralLimitLocation1);
+            mathProperties1.Append(naryLimitLocation1);
+            ThemeFontLanguages themeFontLanguages1 = new ThemeFontLanguages() { Val = "en-US" };
+            ColorSchemeMapping colorSchemeMapping1 = new ColorSchemeMapping() { Background1 = ColorSchemeIndexValues.Light1, Text1 = ColorSchemeIndexValues.Dark1, Background2 = ColorSchemeIndexValues.Light2, Text2 = ColorSchemeIndexValues.Dark2, Accent1 = ColorSchemeIndexValues.Accent1, Accent2 = ColorSchemeIndexValues.Accent2, Accent3 = ColorSchemeIndexValues.Accent3, Accent4 = ColorSchemeIndexValues.Accent4, Accent5 = ColorSchemeIndexValues.Accent5, Accent6 = ColorSchemeIndexValues.Accent6, Hyperlink = ColorSchemeIndexValues.Hyperlink, FollowedHyperlink = ColorSchemeIndexValues.FollowedHyperlink };
+
+            ShapeDefaults shapeDefaults1 = new ShapeDefaults();
+            Ovml.ShapeDefaults shapeDefaults2 = new Ovml.ShapeDefaults() { Extension = V.ExtensionHandlingBehaviorValues.Edit, MaxShapeId = 1026 };
+
+            Ovml.ShapeLayout shapeLayout1 = new Ovml.ShapeLayout() { Extension = V.ExtensionHandlingBehaviorValues.Edit };
+            Ovml.ShapeIdMap shapeIdMap1 = new Ovml.ShapeIdMap() { Extension = V.ExtensionHandlingBehaviorValues.Edit, Data = "1" };
+
+            shapeLayout1.Append(shapeIdMap1);
+
+            shapeDefaults1.Append(shapeDefaults2);
+            shapeDefaults1.Append(shapeLayout1);
+            DecimalSymbol decimalSymbol1 = new DecimalSymbol() { Val = "," };
+            ListSeparator listSeparator1 = new ListSeparator() { Val = ";" };
+            W14.DocumentId documentId1 = new W14.DocumentId() { Val = "7FD518C6" };
+            W15.ChartTrackingRefBased chartTrackingRefBased1 = new W15.ChartTrackingRefBased();
+            W15.PersistentDocumentId persistentDocumentId1 = new W15.PersistentDocumentId() { Val = "{4B1E9AE4-A62B-439B-84E0-F60BAA5072F5}" };
+
+            settings1.Append(zoom1);
+            settings1.Append(proofState1);
+            settings1.Append(defaultTabStop1);
+            settings1.Append(characterSpacingControl1);
+            settings1.Append(compatibility1);
+            settings1.Append(rsids1);
+            settings1.Append(mathProperties1);
+            settings1.Append(themeFontLanguages1);
+            settings1.Append(colorSchemeMapping1);
+            settings1.Append(shapeDefaults1);
+            settings1.Append(decimalSymbol1);
+            settings1.Append(listSeparator1);
+            settings1.Append(documentId1);
+            settings1.Append(chartTrackingRefBased1);
+            settings1.Append(persistentDocumentId1);
+
+            documentSettingsPart1.Settings = settings1;
+        }
+    }
+}
